+++ NEWS.md
@@ -1,5 +1,5 @@
 ## 1.6.1
-  - No changes. Just to resubmit to CRAN.
+  - Minor corrections on documentation and demo files.
 
 ## 1.6.0
   - We are changing the name of the `detectFronts` function to `getGradients`. However, for backwards compatibility, `detectFronts` will still be available until v2.0.0.